+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## dev
+* UNSUPPORTED widget - clients can have different versions of Ed, so it is important to show placeholders for unsupported blocks (#271)
 
 ## 0.21.2 - 2016-09-06